requirements.txt
tensorflow
numpy
pillow
matplotlib
streamlit
